refactor: major refactor of the DebuggerViewModel
