+++ src/jvmMain/kotlin/ui/Controls.kt
@@ -1,6 +1,5 @@
 package ui
 
-import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
 import androidx.compose.material3.Button
@@ -8,27 +7,21 @@ import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.unit.dp
-import viewModel.DebuggerViewModel
+import viewmodel.DebuggerViewModel
 
 @Composable
-fun Controls(viewModel: DebuggerViewModel, setShowFilePicker: (Boolean) -> Unit) {
+fun Controls(viewModel: DebuggerViewModel?, setShowFilePicker: (Boolean) -> Unit) {
     Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
         Button(onClick = { setShowFilePicker(true) }) {
             Text("Open json file")
         }
 
-        OutlinedButton(enabled = viewModel.hasPrevious, onClick = { viewModel.previousEvaluation() }) {
+        OutlinedButton(enabled = viewModel == null || viewModel.hasPrevious, onClick = { viewModel?.previousEvaluation() }) {
             Text("Previous")
         }
 
-        OutlinedButton(enabled = viewModel.hasNext, onClick = { viewModel.nextEvaluation() }) {
+        OutlinedButton(enabled = viewModel == null || viewModel.hasNext, onClick = { viewModel?.nextEvaluation() }) {
             Text("Next")
         }
     }
-}
-
-@Composable
-@Preview
-fun ControlsPreview() {
-    Controls(DebuggerViewModel()) {}
 }
