+++ src/jvmMain/kotlin/viewmodel/DebuggerViewModel.kt
@@ -1,26 +1,26 @@
-package viewModel
+package viewmodel
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import data.BranchTargetRecord
-import data.CodeAttributeRecord
+import data.ExceptionHandlerRecord
 import data.InstructionEvaluationRecord
 import data.StateTracker
 import java.io.File
 
-class DebuggerViewModel {
-    /**
-     * The file that is currently loaded.
-     */
-    var file by mutableStateOf<File?>(null)
-        private set
-
+/**
+ * This view model is a very close representation of the loaded json file.
+ * The view model always belongs to a single file.
+ * When loading a new file, the view model creates and returns a new instance.
+ * @param file The file that is represented by this instance.
+ * @param stateTracker The state tracker that is used to parse the json file.
+ */
+class DebuggerViewModel private constructor(val file: File, stateTracker: StateTracker) {
     /**
      * All the code attributes parsed from the json file.
      */
-    var codeAttributes by mutableStateOf<List<CodeAttributeRecord>>(emptyList())
-        private set
+    val codeAttributes = stateTracker.codeAttributes
 
     /**
      * The current instruction that is being evaluated.
@@ -38,19 +38,31 @@ class DebuggerViewModel {
     )
         private set
 
+    var currentExceptionHandler by mutableStateOf<ExceptionHandlerRecord?>(null)
+        private set
+
     var hasNext by mutableStateOf(false)
         private set
 
     var hasPrevious by mutableStateOf(false)
         private set
 
-    private var stateTracker: StateTracker? = null
     private var currentBlockEvaluation: Int = 0
     private var currentEvaluation: Int = 0
 
-    fun nextEvaluation() {
-        val codeAttributes = stateTracker?.codeAttributes ?: return
+    init {
+        val blockEvaluations = codeAttributes[currentCodeAttribute].blockEvaluations
+        val blockEvaluation = blockEvaluations[currentBlockEvaluation]
+        currentExceptionHandler = blockEvaluation.exceptionHandlerInfo
+        evaluation = blockEvaluation.evaluations[currentEvaluation]
+        currentBlockEvaluationStack = blockEvaluation.branchEvaluationStack ?: emptyList()
+        hasNext = currentCodeAttribute < codeAttributes.size - 1 ||
+            currentBlockEvaluation < blockEvaluations.size - 1 ||
+            currentEvaluation < blockEvaluation.evaluations.size - 1
+        hasPrevious = currentCodeAttribute > 0 || currentBlockEvaluation > 0 || currentEvaluation > 0
+    }
 
+    fun nextEvaluation() {
         val blockEvaluations = codeAttributes[currentCodeAttribute].blockEvaluations
         val evaluations = blockEvaluations[currentBlockEvaluation].evaluations
 
@@ -69,8 +81,6 @@ class DebuggerViewModel {
     }
 
     fun previousEvaluation() {
-        val codeAttributes = stateTracker?.codeAttributes ?: return
-
         if (currentEvaluation > 0) {
             currentEvaluation--
         } else if (currentBlockEvaluation > 0) {
@@ -88,12 +98,11 @@ class DebuggerViewModel {
     }
 
     private fun update() {
-        val codeAttributes = stateTracker?.codeAttributes ?: return
-
         val blockEvaluations = codeAttributes[currentCodeAttribute].blockEvaluations
         val blockEvaluation = blockEvaluations[currentBlockEvaluation]
+        currentExceptionHandler = blockEvaluation.exceptionHandlerInfo
         evaluation = blockEvaluation.evaluations[currentEvaluation]
-        currentBlockEvaluationStack = blockEvaluation.branchEvaluationStack
+        currentBlockEvaluationStack = blockEvaluation.branchEvaluationStack ?: emptyList()
 
         hasNext = currentCodeAttribute < codeAttributes.size - 1 ||
             currentBlockEvaluation < blockEvaluations.size - 1 ||
@@ -103,37 +112,28 @@ class DebuggerViewModel {
     }
 
     /**
-     * Loads the json file at the given path into the view model.
+     * Loads the json file at the given path and returns a new view model.
      */
-    fun loadJson(path: String) {
-        reset()
-        file = File(path)
-        try {
-            stateTracker = StateTracker.fromJson(path)
+    fun loadJson(path: String): DebuggerViewModel {
+        val file = File(path)
+        return try {
+            val stateTracker = StateTracker.fromJson(path)
+            DebuggerViewModel(file, stateTracker)
         } catch (e: Exception) {
             println("Error while parsing json file: $e")
-            return
+            this
         }
-        codeAttributes = stateTracker?.codeAttributes ?: emptyList()
-        update()
     }
 
-    /**
-     * Resets the view model to its initial state.
-     */
-    fun reset() {
-        hasNext = false
-        hasPrevious = false
-
-        currentEvaluation = 0
-        currentBlockEvaluation = 0
-        currentCodeAttribute = 0
-
-        currentBlockEvaluationStack = emptyList()
-        codeAttributes = emptyList()
-
-        evaluation = null
-        stateTracker = null
-        file = null
+    companion object {
+        /**
+         * Create a new view model given a path to a json file.
+         * Used to create the initial view model.
+         */
+        fun fromJson(path: String): DebuggerViewModel {
+            val file = File(path)
+            val stateTracker = StateTracker.fromJson(path)
+            return DebuggerViewModel(file, stateTracker)
+        }
     }
 }
